Refactor: Move post component logic to hooks usePost and useUser
