+++ src/components/Post/styles.ts
